feat: implement comment deletion functionality; add deleteComment API mutation and integrate with CommentCard for user actions, also fixed cascading in comment and replies entities

+++ backend/project/src/main/java/com/instragram/project/service/CommentService.java
@@ -12,6 +12,7 @@ import com.instragram.project.model.Comment;
 import com.instragram.project.repository.AppUserRepository;
 import com.instragram.project.repository.CommentRepository;
 
+import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 
 @Service
@@ -44,10 +45,15 @@ public class CommentService {
    }
 
    // Delete the comment
-   public void deleteComment(Long commentId) {
-
+   @Transactional
+   public void deleteComment(Long commentId, String username) {
+      AppUser appUser = appUserRepository.findByUsername(username);
       Comment comment = commentRepository.findById(commentId)
             .orElseThrow(() -> new RuntimeException("Comment not found with id: " + commentId));
+
+      if (appUser != comment.getAppUser() && appUser != comment.getPost().getAppUser()) {
+         throw new RuntimeException("You do not have permission to delete this comment");
+      }
 
       commentRepository.delete(comment);
    }
